refactor: Extract shared declaration from test_query.
